test_six.py: test_MAXSIZE allow ValueError or OverflowError
because the previous code was expecting ValueError on Python 2.4,
but I was getting OverflowError on Python 2.4.4 on OS X but ValueError on
Python 2.4.6 on Ubuntu.

So perhaps knowing which exception will be raised is not that straightforward,
so I made it check for either one.

--HG--
branch : test_MAXSIZE__ValueError_or_OverflowError